added Auth controller

+++ src/main/java/pl/travel/travelapp/services/UserService.java
@@ -33,6 +33,7 @@ import pl.travel.travelapp.repositories.UserRepository;
 
 import java.io.File;
 import java.io.FileNotFoundException;
+import java.security.Principal;
 import java.time.LocalDate;
 import java.util.HashMap;
 import java.util.Map;
@@ -76,25 +77,47 @@ public class UserService implements UserDetailsService, UserServiceInterface {
     //Checking the data and returning the registration result
     @Override
     public ResponseEntity<String> userRegister(UserRegisterDTO user) {
-        if (userRepository.checkIfExist(user.getLogin(), user.getEmail()).isPresent())
-            return new ResponseEntity<>("an account with this login or e-mail address already exists", HttpStatus.CONFLICT);
-        if (!isValidPassword(user.getPassword(), regex))
-            return new ResponseEntity<>("Your password doesn't suit requirements", HttpStatus.NOT_ACCEPTABLE);
-        if (!countryRepository.findFirstByCountry(user.getNationality()).isPresent())
-            return new ResponseEntity<>("Country doesn't exist", HttpStatus.NOT_FOUND);
         try {
-            boolean isCreated = userRegisterSave(user);
-            if (isCreated && mailService.sendMailByGoogleMailApi(user.getEmail(), "Travel App Account", HtmlContent.readHtmlRegistration(user.getLogin(), ""))) {
-                return new ResponseEntity<>("Account has been created", HttpStatus.OK);
+
+            if(checkIfDataIsNotBlank(user)){
+                if (userRepository.checkIfExist(user.getLogin(), user.getEmail()).isPresent())
+                    return new ResponseEntity<>("an account with this login or e-mail address already exists", HttpStatus.CONFLICT);
+                if (!isValidPassword(user.getPassword(), regex))
+                    return new ResponseEntity<>("Your password doesn't suit requirements", HttpStatus.NOT_ACCEPTABLE);
+                if (!countryRepository.findFirstByCountry(user.getNationality()).isPresent())
+                    return new ResponseEntity<>("Country doesn't exist", HttpStatus.NOT_FOUND);
+                try {
+                    boolean isCreated = userRegisterSave(user);
+                    if (isCreated && mailService.sendMailByGoogleMailApi(user.getEmail(), "Travel App Account", HtmlContent.readHtmlRegistration(user.getLogin(), ""))) {
+                        return new ResponseEntity<>("Account has been created", HttpStatus.OK);
+                    }
+                } catch (Exception e) {
+                    e.printStackTrace();
+                    return new ResponseEntity<>("Account hasn't been created", HttpStatus.NOT_IMPLEMENTED);
+                }
+
             }
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new ResponseEntity<>("Account hasn't been created", HttpStatus.NOT_IMPLEMENTED);
+        }catch (NullPointerException e){
+            return new ResponseEntity<>("Send blank data is forbidden", HttpStatus.NOT_ACCEPTABLE);
         }
         return new ResponseEntity<>("Undefined error", HttpStatus.NOT_FOUND);
     }
 
-    @Override
+
+    //Checking whether user data is not blank, if yes the method throws NullPointerException
+    private boolean checkIfDataIsNotBlank(UserRegisterDTO user) throws NullPointerException{
+        if(user.getEmail().isBlank()) throw new NullPointerException();
+        if(user.getPassword().isBlank()) throw new NullPointerException();
+        if(user.getLogin().isBlank()) throw new NullPointerException();
+        if(user.getNationality().isBlank()) throw new NullPointerException();
+        if(user.getBirthDay() == null) throw new NullPointerException();
+        if(user.getFirstName().isBlank()) throw new NullPointerException();
+        if(user.getSurName().isBlank()) throw new NullPointerException();
+        return true;
+    }
+
+
+
     @Transactional
     public boolean userRegisterSave(UserRegisterDTO user) throws Exception {
         try {
@@ -113,52 +136,61 @@ public class UserService implements UserDetailsService, UserServiceInterface {
             return true;
         } catch (Exception e) {
             System.err.println("Account hasn't been created " + e);
+            return false;
         }
-        return false;
     }
 
-    //Saving new user to database
-
+    //Set user account enable
     @Override
     @Transactional
     public boolean enableUserAccount(String token) {
-        JWTVerifier jwtVerifier = JWT.require(Algorithm.HMAC512(key)).build();
-        DecodedJWT verify = jwtVerifier.verify(token);
-        String username = verify.getClaim("username").asString();
-        Optional<User> userApp = userRepository.findFirstByLogin(username);
-        if (userApp.isEmpty()) return false;
-        userApp.ifPresent(userAppUpdate -> {
-            userAppUpdate.setEnable(true);
-            userRepository.save(userAppUpdate);
-        });
-        return true;
+        try {
+            JWTVerifier jwtVerifier = JWT.require(Algorithm.HMAC512(key)).build();
+            DecodedJWT verify = jwtVerifier.verify(token);
+            String username = verify.getClaim("username").asString();
+            Optional<User> userApp = userRepository.findFirstByLogin(username);
+            if (userApp.isEmpty()) return false;
+            userApp.ifPresent(userAppUpdate -> {
+                userAppUpdate.setEnable(true);
+                userRepository.save(userAppUpdate);
+            });
+            return true;
+        }catch (Exception e){
+            return false;
+        }
     }
 
     @Override
     @Transactional
     public ResponseEntity<String> changePassword(Map<String, String> fields) {
 
-        String password = fields.get("password");
-        String token = fields.get("token");
         try {
-            JWTVerifier jwtVerifier = JWT.require(Algorithm.HMAC512(key)).build();
-            DecodedJWT verify = jwtVerifier.verify(token);
-            String login = verify.getClaim("username").asString();
-            Optional<User> userApp = userRepository.findFirstByLogin(login);
-            if (userApp.isPresent()) {
-                User userToSave = userApp.get();
-                userToSave.setPassword(passwordEncoder.encode(password));
-                try {
-                    userRepository.save(userToSave);
-                    return new ResponseEntity<>("Password has been changed", HttpStatus.OK);
-                } catch (Exception e) {
-                    System.err.println(e);
-                    return new ResponseEntity("Password hasn't been changed", HttpStatus.NOT_MODIFIED);
+            String password = fields.get("password");
+            String token = fields.get("token");
+
+
+            try {
+                JWTVerifier jwtVerifier = JWT.require(Algorithm.HMAC512(key)).build();
+                DecodedJWT verify = jwtVerifier.verify(token);
+                String login = verify.getClaim("username").asString();
+                Optional<User> userApp = userRepository.findFirstByLogin(login);
+                if (userApp.isPresent()) {
+                    User userToSave = userApp.get();
+                    userToSave.setPassword(passwordEncoder.encode(password));
+                    try {
+                        userRepository.save(userToSave);
+                        return new ResponseEntity<>("Password has been changed", HttpStatus.OK);
+                    } catch (Exception e) {
+                        System.err.println(e);
+                        return new ResponseEntity("Password hasn't been changed", HttpStatus.NOT_MODIFIED);
+                    }
                 }
+                return new ResponseEntity(HttpStatus.NOT_FOUND);
+            } catch (UsernameNotFoundException e) {
+                return new ResponseEntity(HttpStatus.NOT_FOUND);
             }
-            return new ResponseEntity(HttpStatus.NOT_FOUND);
-        } catch (UsernameNotFoundException e) {
-            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        }catch (NullPointerException e){
+            return new ResponseEntity(HttpStatus.FORBIDDEN);
         }
     }
 
@@ -166,10 +198,15 @@ public class UserService implements UserDetailsService, UserServiceInterface {
     //Send mail which allow to change password
     @Override
     public boolean forgetPassword(String email) {
-        Optional<User> user = userRepository.findFirstByEmail(email);
-        if (user.isPresent()) {
-            mailService.sendMailByGoogleMailApi(email, "Travel App, new password", HtmlContent.newPasswordHtml(user.get().getLogin(), ""));
-            return true;
+        try {
+            if(email.isBlank()) return false;
+            Optional<User> user = userRepository.findFirstByEmail(email);
+            if (user.isPresent()) {
+                mailService.sendMailByGoogleMailApi(email, "Travel App, new password", HtmlContent.newPasswordHtml(user.get().getLogin(), ""));
+                return true;
+            }
+        }catch (NullPointerException e){
+            return false;
         }
         return false;
     }
@@ -187,36 +224,43 @@ public class UserService implements UserDetailsService, UserServiceInterface {
                 return new ResponseEntity<>(user, HttpStatus.OK);
             }
             return new ResponseEntity(HttpStatus.NOT_FOUND);
-        } catch (UsernameNotFoundException e) {
+        } catch (UsernameNotFoundException | NullPointerException e) {
             return new ResponseEntity(HttpStatus.NOT_FOUND);
         }
     }
     @Override
-    public ResponseEntity deleteAccountMessage(UserDetails user) {
-        User userToDelete = userRepository.findFirstByLogin(user.getUsername()).get();
-        mailService.sendMailByGoogleMailApi(userToDelete.getEmail(), "Delete account, Travel App", HtmlContent.deleteAccountTemplate(user.getUsername(), ""));
+    public ResponseEntity deleteAccountMessage(Principal user) {
+        User userToDelete = userRepository.findFirstByLogin(user.getName()).get();
+        mailService.sendMailByGoogleMailApi(userToDelete.getEmail(), "Delete account, Travel App", HtmlContent.deleteAccountTemplate(user.getName(), ""));
         return new ResponseEntity(HttpStatus.OK);
     }
 
     @Override
     @Transactional
-    public ResponseEntity<String> deleteAccount(UserDetails user, String password) {
-        Optional<User> userApp = userRepository.findFirstByLogin(user.getUsername());
-        if (userApp.isPresent()) {
-            User userToDelete = userApp.get();
-            if (passwordEncoder.matches(password,userToDelete.getPassword())) {
-                try {
-                    userRepository.delete(userToDelete);
-                    return new ResponseEntity<>("Account has been deleted", HttpStatus.OK);
-                } catch (Exception e) {
-                    System.err.println(e);
-                    return new ResponseEntity("Account hasn't been deleted", HttpStatus.NOT_MODIFIED);
+    public ResponseEntity<String> deleteAccount(Principal user,Map<String, String> fields) {
+        try {
+            String password = fields.get("password");
+            if(password.isBlank()) throw new NullPointerException();
+            Optional<User> userApp = userRepository.findFirstByLogin(user.getName());
+            if (userApp.isPresent()) {
+                User userToDelete = userApp.get();
+                if (passwordEncoder.matches(password,userToDelete.getPassword())) {
+                    try {
+                        userRepository.delete(userToDelete);
+                        return new ResponseEntity<>("Account has been deleted", HttpStatus.OK);
+                    } catch (Exception e) {
+                        System.err.println(e);
+                        return new ResponseEntity("Account hasn't been deleted", HttpStatus.NOT_MODIFIED);
+                    }
+                } else {
+                    return new ResponseEntity<>("PASSWORD INCORRECT", HttpStatus.NOT_ACCEPTABLE);
                 }
-            } else {
-                return new ResponseEntity<>("PASSWORD INCORRECT", HttpStatus.NOT_ACCEPTABLE);
-            }
 
+            }
+        }catch (NullPointerException e){
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
         }
+
         return new ResponseEntity(HttpStatus.NOT_FOUND);
     }
 
@@ -236,6 +280,9 @@ public class UserService implements UserDetailsService, UserServiceInterface {
     //          Class to test
     @EventListener(ApplicationReadyEvent.class)
     public void test() {
+
+
+
 //        UserRegisterDTO userRegisterDTO = new UserRegisterDTO();
 //        userRegisterDTO.setEmail("faronnorbertkrk@gmail.com");
 //        userRegisterDTO.setBirthDay(LocalDate.now());
@@ -252,7 +299,6 @@ public class UserService implements UserDetailsService, UserServiceInterface {
 //        fields.put("token",userLoginDTO.getToken());
 //        System.out.println(changePassword(fields));
 //        forgetPassword("faronnorbertkrk@gmail.com");
-//        UserDetails user = userRepository.findByLogin("norbert1517");
 //       deleteAccount(user, "N@jwalxcm123ka");
     }
